+++ bundler.js
@@ -7,6 +7,8 @@ var fs = require("fs");
 var Less = require('less');
 var UglifyJS = require("uglify-js");
 var filesJson = require("./src/files");
+var crypto = require('crypto');
+var babelCore = require("babel-core");
 
 const EOL = require("os").EOL;
 const PATH_SEP = require("path").sep;
@@ -28,7 +30,12 @@ const _CONST = {
 	LESS_FILE: "less/onix.less",
 	CSS_FILE: "dist/onix.css",
 	LESS_TEST_FILE: "test/less/main.less",
-	CSS_TEST_FILE: "test/css/main.css"
+	CSS_TEST_FILE: "test/css/main.css",
+	CACHE_FILE: "./cache.json",
+	ES6_ENABLE: false,
+	ES6_PRESETS: [
+		"babel-preset-es2015"
+	]
 };
 
 const _JS_MINIMAL_FILES = filesJson.minimal.map((item) => {
@@ -36,6 +43,10 @@ const _JS_MINIMAL_FILES = filesJson.minimal.map((item) => {
 });
 
 const _JS_FILES = filesJson.onix.map((item) => {
+	return _CONST.JS_SRC_PATH + item;
+});
+
+const _ES6_SKIP = filesJson.es6_skip.map((item) => {
 	return _CONST.JS_SRC_PATH + item;
 });
 
@@ -81,7 +92,7 @@ class Common {
 	 * @param  {Number} size
 	 * @return {String}
 	 */
-	static humanLength(size) {
+	static formatSize(size) {
 		size = size || 0;
 
 		let sizeKB = size / 1024;
@@ -114,6 +125,33 @@ class Common {
 	}
 
 	/**
+	 * Remove file.
+	 * @param  {String} name
+	 * @return {Promise}
+	 */
+	static removeFile(name) {
+		return new Promise((resolve, reject) => {
+			name = name || "";
+
+			fs.stat(name, (err, stats) => {
+				if (stats) {
+					fs.unlink(name, (err) => {
+						if (err) {
+							reject(err);
+						} 
+						else {
+							resolve();
+						}
+					});
+				}
+				else {
+					resolve();
+				}
+			});
+		});
+	}
+
+	/**
 	 * Write file.
 	 * @param  {String} name
 	 * @param  {String} data
@@ -121,6 +159,8 @@ class Common {
 	 */
 	static writeFile(name, data) {
 		return new Promise((resolve, reject) => {
+			name = name || "";
+
 			fs.writeFile(name, data, 'utf8', (err) => {
 				if (err) {
 					reject(err);
@@ -280,23 +320,193 @@ class Watcher {
 	}
 };
 
+class TimeStop {
+	constructor() {
+		this._cache = {};
+	}
+
+	start(label) {
+		let ct = Date.now();
+		label = label || ct;
+		this._cache[label] = ct;
+	}
+
+	end(label) {
+		label = label || "";
+		let ct = Date.now();
+		let cacheItem = this._cache[label];
+
+		if (cacheItem) {
+			let diff = ct - cacheItem;
+
+			if (diff < 1000) {
+				return diff + " ms";
+			}
+			else {
+				return (diff / 1000).toFixed(2) + " s";
+			}
+		}
+		else return "0 ms";
+	}
+};
+
+class ES6 {
+	constructor() {
+		this._fileName = _CONST.CACHE_FILE;
+		this._hash = {};
+		this._cache = fs.existsSync(this._fileName) ? require(this._fileName) : {};
+	}
+
+	/**
+	 * Transpile data
+	 * 
+	 * @param  {String} path
+	 * @param  {String} data
+	 * @return {Object} { code: "", cache: ... }
+	 */
+	transpile(path, data) {
+		let output = {
+			code: "",
+			cache: false
+		};
+
+		let md5hash = this.md5hash(data);
+
+		// cache record
+		this._hash[md5hash] = true;
+
+		// cache?
+		if (!(md5hash in this._cache)) {
+			let presets = _CONST.ES6_PRESETS;
+
+			let es6config = {
+				presets: presets.map((preset) => {
+					return require.resolve(preset)
+				})
+			};
+
+			try {
+				let transform = babelCore.transform(data, es6config);
+
+				this._cache[md5hash] = transform.code;
+			}
+			catch (err) {
+				this.parseES6error(path, data, err);
+			}
+		}
+		else {
+			output.cache = true;
+		}
+
+		output.code = this._cache[md5hash] || "";
+
+		return output;
+	}
+
+	md5hash(data) {
+		return crypto.createHash('md5').update(data).digest('hex');
+	}
+
+	/**
+	 * Parse ES6 transform error. Show path and lineNumber + line content.
+	 * Output is shown in console.
+	 * @param  {String} path
+	 * @param  {String} data
+	 * @param  {Object} err 
+	 */
+	parseES6error(path, data, err) {
+		Common.col("ES6 transform error {0} {1}", path, err.message);
+
+		let matchLine = err.message.match(/\([0-9]+:[0-9]\)/);
+
+		if (matchLine) {
+			let lineNumber = matchLine[0].replace(/\(|\)/g, "");
+			lineNumber = parseInt(lineNumber.split(":")[0], 10);
+
+			let lines = data.split("\n");
+
+			if (lineNumber >= 0 && lineNumber < lines.length) {
+				Common.col("Line {1}: {0}", lines[lineNumber].trim(), lineNumber);
+			}
+		}
+	}
+
+	saveCache() {
+		return new Promise((resolve, reject) => {
+			var hashKeys = Object.keys(this._hash);
+
+			if (!hashKeys.length) {
+				resolve();
+			}
+			else {
+				// remove unused keys
+				Object.keys(this._cache).forEach((key) => {
+					if (hashKeys.indexOf(key) == -1) {
+						delete this._cache[key];
+					}
+				}, this);
+
+				// write cache
+				let data = JSON.stringify(this._cache, null, 4);
+
+				Common.writeFile(this._fileName, data).then(() => {
+					Common.col("Cache file {0} {1} was written!", this._fileName, Common.formatSize(data.length));
+					resolve();
+				}, () => {
+					Common.col("Cache file {0} write file error!", this._fileName);
+					reject();
+				});
+			}
+		});
+	}
+
+	clearCache() {
+		return new Promise((resolve, reject) => {
+			Common.removeFile(this._fileName).then(() => {
+				Common.col("Cache file {0} was removed!", this._fileName);
+				resolve();
+			}, (err) => {
+				Common.col("Cache file {0} remove file error: {1}!", this._fileName, err.message);
+				reject();
+			});
+		});
+	}
+}
+
 class Bundler {
 	constructor() {
 		this._arg = this._getArg();
 
+		this._es6 = new ES6();
+
 		this._setupWatcher();
+
+		this._timeStop = new TimeStop();
 
 		this._filesCache = {};
 
-		this._loadFiles().then(() => {
-			Common.col("Bundler help");
-			Common.col("------------");
-			Common.col("./bundler.js watch - watch and compile js & less files");
-			Common.col("./bundler.js dev - compile for dev");
-			Common.col("./bundler.js dist - compile for dist");
-			Common.col("./bundler.js doc - documentation");
-			Common.col("./bundler.js - default is watch option");
+		Common.col("Bundler help");
+		Common.col("------------");
+		Common.col("./bundler.js watch - watch and compile js & less files");
+		Common.col("./bundler.js dev - compile for dev");
+		Common.col("./bundler.js dist - compile for dist");
+		Common.col("./bundler.js doc - documentation");
+		Common.col("./bundler.js clear - clear ES6 cache");
+		Common.col("./bundler.js - default is watch option");
 
+		// pre files operations
+		switch (this._arg) {
+			case "doc":
+				this._jsduck();
+				return;
+
+			case "clear":
+				this._clear();
+				return;
+		}
+
+		// post files operations
+		this._loadFiles().then(() => {
 			switch (this._arg) {
 				case "watch":
 					this._dev("watcher");
@@ -308,10 +518,6 @@ class Bundler {
 
 				case "dist":
 					this._dist();
-					break;
-
-				case "doc":
-					this._jsduck();
 					break;
 
 				default:
@@ -327,11 +533,27 @@ class Bundler {
 		return new Promise((resolve, reject) => {
 			let c = _CONST;
 			let allFiles = [c.HEADER_FILE].concat(_JS_FILES);
+			let es6File = 0;
+			let es6FileCache = 0;
+
+			if (_CONST.ES6_ENABLE) {
+				Common.col("Starting ES6 transpile...");
+
+				this._timeStop.start("ES6");
+			}
 
 			Common.readFiles(allFiles).then((all) => {
 				all.forEach((file) => {
-					this._filesCache[file.path] = file.data;
+					let o = this._setFileCache(file.path, file.data);
+
+					es6File += o.es6File ? 1 : 0;
+					es6FileCache += o.es6FileCache ? 1 : 0;
 				});
+
+				if (_CONST.ES6_ENABLE) {
+					Common.col("ES6 transpilation: {0} files, {1} cached, takes {2}",
+								es6File, es6FileCache, this._timeStop.end("ES6"));
+				}
 
 				resolve();
 			}, (err) => {
@@ -356,10 +578,11 @@ class Bundler {
 
 				if (_JS_FILES.indexOf(path) != -1) {
 					Common.readFile(path).then((data) => {
-						// todo update onix.js
-						this._filesCache[path] = data;
+						this._setFileCache(path, data);
 
-						this._makeJS(_JS_FILES, _CONST.JS_OUTPUT);
+						this._makeJS(_JS_FILES, _CONST.JS_OUTPUT).then(() => {
+							this._makeCache();
+						});
 					});
 				}
 				else if (path.indexOf(_CONST.WATCH_PATHS[2]) != -1) {
@@ -387,6 +610,41 @@ class Bundler {
 		else return "";
 	}
 
+	/**
+	 * Set file cache.
+	 * 
+	 * @param {String} path
+	 * @param {String} data
+	 * @return {Object} ES6 info
+	 */
+	_setFileCache(path, data) {
+		let output = {
+			es6File: false,
+			es6FileCache: false
+		};
+
+		let outputData = data;
+
+		// es6
+		if (_CONST.ES6_ENABLE) {
+			if (path.indexOf(".js") == -1 || _ES6_SKIP.indexOf(path) != -1) {
+				Common.col("ES6 skip path {0}", path);
+			}
+			else {
+				let transpile = this._es6.transpile(path, data);
+
+				output.es6File = true;
+				output.es6FileCache = transpile.cache;
+
+				outputData = transpile.code;
+			}
+		}
+
+		this._filesCache[path] = outputData;
+
+		return output;
+	}
+
 	_getJSfromCache(path) {
 		let data = this._filesCache[path] || "";
 		let onixPath = this._getOnixJSPath();
@@ -406,11 +664,18 @@ class Bundler {
 				output.push(this._getJSfromCache(file));
 			});
 
+			// output
+			let outputStr = output.join(EOL);
+
+			if (_CONST.ES6_ENABLE) {
+				outputStr = outputStr.replace(/"use strict";/g, "");
+			}
+			
 			// clear blank lines
-			let outputStr = output.join(EOL).replace(/(^[ \t]*\n)/gm, "");
+			outputStr = outputStr.replace(/(^[ \t]*\n)/gm, "");
 
 			Common.writeFile(outputFile, outputStr).then(() => {
-				Common.col("Write JS {0} {1}", outputFile, Common.humanLength(outputStr.length));
+				Common.col("Write JS {0} {1}", outputFile, Common.formatSize(outputStr.length));
 				resolve();
 			}, () => {
 				Common.col("Write JS {0} write file error!", e.filename);
@@ -444,7 +709,7 @@ class Bundler {
 					else {
 						// write output
 						Common.writeFile(outputFile, output.css).then(() => {
-							Common.col("Write CSS {0} {1}", outputFile, Common.humanLength(output.css.length));
+							Common.col("Write CSS {0} {1}", outputFile, Common.formatSize(output.css.length));
 							resolve();
 						}, () => {
 							Common.col("Write {0} write file error!", e.filename);
@@ -457,6 +722,10 @@ class Bundler {
 				reject();
 			});
 		});
+	}
+
+	_makeCache() {
+		return this._es6.saveCache();
 	}
 	
 	/**
@@ -472,7 +741,7 @@ class Bundler {
 
 			// write output
 			Common.writeFile(outputFile, data).then(() => {
-				Common.col("Write JS {0} {1}", outputFile, Common.humanLength(result.code.length));
+				Common.col("Write JS {0} {1}", outputFile, Common.formatSize(result.code.length));
 				resolve();
 			}, () => {
 				Common.col("Write {0} write file error!", e.filename);
@@ -573,6 +842,8 @@ class Bundler {
 	 */
 	_jsduck() {
 		return new Promise((resolve, reject) => {
+			Common.col("JSDuck output:");
+
 			exec("jsduck src/core/*.js src/onix/*.js src/utils/*.js --output docs --warnings=-nodoc,-dup_member,-link_ambiguous --external=XMLHttpRequest,$q", (error, stdout, stderr) => {
 				if (stdout) {
 					console.log(stdout);
@@ -604,6 +875,9 @@ class Bundler {
 		}, {
 			method: "_makeLess",
 			args: [c.LESS_TEST_FILE, c.CSS_TEST_FILE, false],
+			scope: this
+		}, {
+			method: "_makeCache",
 			scope: this
 		}]).then(() => {
 			let pd = this._parseOnixJSFile();
@@ -653,6 +927,11 @@ class Bundler {
 		}]).then(() => {
 			console.log("Dist is done");
 		});
+	}
+
+	// clear es6 cache
+	_clear() {
+		this._es6.clearCache();
 	}
 }